Fixed page freeze when loading directly into agenda item page
